Refactor PbF/PbEM Poster code (#4454)
* Rewire a lot of code part 1

* Rewire a lot of code part 2

* Rewire a lot of code part 3

* Rewire a lot of code part 4

* Don't ignore future return value

* Fix checkstyleMain

* Add missing javadoc

* Fix javadoc

* Remove IValidator interface

* Add correct checks back

* Fix Dice Server Settings

* Fix static class

* Fix checkstyle

* Fix SetupPanel loading

* Fix exceptions

* Fix false populating

* Fix simple issues

* Use optional

* Fix potential NullPointerExceptions

* Move logic to IForumPoster to decrease implementation visibility

* Add Util#isInt test case

* Move method and add Unit test

* Close InputStream

* Add trivial test

* Enhance use of optional

* Use different utility method

* Fix checkstyle

* Workaround explicit dependency

* Fix if statement

* Rename field to be more accurate

* Suppress warnings

+++ game-core/src/main/java/games/strategy/engine/pbem/PbemMessagePoster.java
@@ -2,7 +2,11 @@ package games.strategy.engine.pbem;
 
 import java.io.File;
 import java.io.IOException;
+import java.io.ObjectInputStream;
+import java.io.ObjectOutputStream;
 import java.io.Serializable;
+import java.util.Optional;
+import java.util.concurrent.Future;
 import java.util.logging.Level;
 
 import javax.swing.JComponent;
@@ -11,6 +15,7 @@ import javax.swing.SwingUtilities;
 
 import games.strategy.engine.data.GameData;
 import games.strategy.engine.data.PlayerId;
+import games.strategy.engine.data.properties.GameProperties;
 import games.strategy.engine.framework.GameDataFileUtils;
 import games.strategy.engine.history.IDelegateHistoryWriter;
 import games.strategy.triplea.delegate.remote.IAbstractForumPosterDelegate;
@@ -22,69 +27,58 @@ import lombok.extern.java.Log;
 /**
  * This class is responsible for posting turn summary and email at the end of each round in a PBEM game.
  * A new instance is created at end of turn, based on the Email and a forum poster stored in the game data.
- * The needs to be serialized since it is invoked through the IAbstractEndTurnDelegate which require all objects to be
- * serializable although the PBEM games will always be local
+ * This class does only implement {@link Serializable} because otherwise the delegate would reject it,
+ * even though this class is for local use only.
  */
 @Log
 public class PbemMessagePoster implements Serializable {
-  public static final String FORUM_POSTER_PROP_NAME = "games.strategy.engine.pbem.IForumPoster";
-  public static final String EMAIL_SENDER_PROP_NAME = "games.strategy.engine.pbem.IEmailSender";
-  public static final String PBEM_GAME_PROP_NAME = "games.strategy.engine.pbem.PbemMessagePoster";
-  private static final long serialVersionUID = 2256265436928530566L;
-
-  private final IForumPoster forumPoster;
-  private final IEmailSender emailSender;
-  private transient File saveGameFile = null;
-  private transient String turnSummary = null;
-  private final transient String saveGameRef = null;
-  private transient String turnSummaryRef = null;
-  private transient String emailSendStatus;
-  private final transient PlayerId currentPlayer;
-  private final transient int roundNumber;
-  private final transient String gameNameAndInfo;
+  private static final long serialVersionUID = -1L;
+  private final GameProperties gameProperties;
+  private File saveGameFile = null;
+  private String turnSummary = null;
+  private String turnSummaryRef = null;
+  private String emailSendStatus;
+  private final PlayerId currentPlayer;
+  private final int roundNumber;
+  private final String gameNameAndInfo;
 
   public PbemMessagePoster(final GameData gameData, final PlayerId currentPlayer, final int roundNumber,
       final String title) {
     this.currentPlayer = currentPlayer;
     this.roundNumber = roundNumber;
-    forumPoster = (IForumPoster) gameData.getProperties().get(FORUM_POSTER_PROP_NAME);
-    emailSender = (IEmailSender) gameData.getProperties().get(EMAIL_SENDER_PROP_NAME);
+    gameProperties = gameData.getProperties();
     gameNameAndInfo =
         "TripleA " + title + " for game: " + gameData.getGameName() + ", version: " + gameData.getGameVersion();
   }
 
   public boolean hasMessengers() {
-    return forumPoster != null || emailSender != null;
+    return gameProperties.get(IForumPoster.NAME) != null || gameProperties.get(IEmailSender.SUBJECT) != null;
   }
 
   public static boolean gameDataHasPlayByEmailOrForumMessengers(final GameData gameData) {
-    if (gameData == null) {
-      return false;
-    }
-    final IForumPoster forumPoster = (IForumPoster) gameData.getProperties().get(FORUM_POSTER_PROP_NAME);
-    final IEmailSender emailSender = (IEmailSender) gameData.getProperties().get(EMAIL_SENDER_PROP_NAME);
-    final boolean isPbem = gameData.getProperties().get(PBEM_GAME_PROP_NAME, false);
-    return isPbem && (forumPoster != null || emailSender != null);
-  }
-
-  public IForumPoster getForumPoster() {
-    return forumPoster;
-  }
-
-  public void setTurnSummary(final String turnSummary) {
-    this.turnSummary = turnSummary;
+    return gameData != null
+        && (gameData.getProperties().get(IForumPoster.NAME) != null
+        || gameData.getProperties().get(IEmailSender.SUBJECT) != null);
   }
 
   public void setSaveGame(final File saveGameFile) {
     this.saveGameFile = saveGameFile;
   }
 
-  public String getTurnSummaryRef() {
-    return turnSummaryRef;
+  private Optional<IForumPoster> newForumPoster() {
+    final String name = gameProperties.get(IForumPoster.NAME, "");
+    if (name.isEmpty()) {
+      return Optional.empty();
+    }
+    return Optional.of(IForumPoster.newInstanceByName(name, gameProperties.get(IForumPoster.TOPIC_ID, 0)));
   }
 
-  public String getSaveGameRef() {
-    return saveGameRef;
+  private Optional<IEmailSender> newEmailSender() {
+    final String subject = gameProperties.get(IEmailSender.SUBJECT, "");
+    if (subject.isEmpty()) {
+      return Optional.empty();
+    }
+    return Optional.of(IEmailSender.newInstance(subject, gameProperties.get(IEmailSender.RECIPIENTS, "")));
   }
 
   /**
@@ -93,23 +87,22 @@ public class PbemMessagePoster implements Serializable {
    * @param historyWriter the history writer (which has no effect since save game has already be generated) // todo (kg)
    * @return true if all posts were successful
    */
-  public boolean post(final IDelegateHistoryWriter historyWriter, final String title, final boolean includeSaveGame) {
-    boolean forumSuccess = true;
+  public boolean post(final IDelegateHistoryWriter historyWriter, final String title) {
+    final Optional<IForumPoster> forumPoster = newForumPoster();
+
     final StringBuilder saveGameSb = new StringBuilder().append("triplea_");
-    if (forumPoster != null) {
-      saveGameSb.append(forumPoster.getTopicId()).append("_");
+    if (forumPoster.isPresent()) {
+      saveGameSb.append(gameProperties.get(IForumPoster.TOPIC_ID)).append("_");
     }
     saveGameSb.append(currentPlayer.getName(), 0, Math.min(3, currentPlayer.getName().length() - 1))
         .append(roundNumber);
     final String saveGameName = GameDataFileUtils.addExtension(saveGameSb.toString());
-    if (forumPoster != null) {
-      if (includeSaveGame) {
-        forumPoster.addSaveGame(saveGameFile, saveGameName);
-      }
+    Future<String> forumSuccess = null;
+    if (forumPoster.isPresent()) {
       try {
-        forumSuccess = forumPoster.postTurnSummary((gameNameAndInfo + "\n\n" + turnSummary),
-            "TripleA " + title + ": " + currentPlayer.getName() + " round " + roundNumber);
-        turnSummaryRef = forumPoster.getTurnSummaryRef();
+        forumSuccess = forumPoster.get().postTurnSummary((gameNameAndInfo + "\n\n" + turnSummary),
+            "TripleA " + title + ": " + currentPlayer.getName() + " round " + roundNumber, saveGameFile.toPath());
+        turnSummaryRef = forumSuccess.get();
         if (turnSummaryRef != null && historyWriter != null) {
           historyWriter.startEvent("Turn Summary: " + turnSummaryRef);
         }
@@ -117,37 +110,32 @@ public class PbemMessagePoster implements Serializable {
         log.log(Level.SEVERE, "Failed to post game to forum", e);
       }
     }
-    boolean emailSuccess = true;
-    if (emailSender != null) {
-      final StringBuilder subjectPostFix = new StringBuilder(currentPlayer.getName());
-      subjectPostFix.append(" - ").append("round ").append(roundNumber);
+    final Optional<IEmailSender> emailSender = newEmailSender();
+    final boolean emailSuccess = emailSender.map(sender -> {
       try {
-        emailSender.sendEmail(subjectPostFix.toString(), convertToHtml((gameNameAndInfo + "\n\n" + turnSummary)),
-            saveGameFile, saveGameName);
-        emailSendStatus = "Success, sent to " + emailSender.getToAddress();
+        sender.sendEmail(currentPlayer.getName() + " - round " + roundNumber,
+            convertToHtml((gameNameAndInfo + "\n\n" + turnSummary)), saveGameFile, saveGameName);
+        emailSendStatus = "Success, sent to " + gameProperties.get(IEmailSender.RECIPIENTS);
+        return true;
       } catch (final IOException e) {
-        emailSuccess = false;
         emailSendStatus = "Failed! Error " + e.getMessage();
         log.log(Level.SEVERE, "Failed to send game via email", e);
+        return false;
       }
-    }
+    }).orElse(false);
     if (historyWriter != null) {
       final StringBuilder sb = new StringBuilder("Post Turn Summary");
-      if (forumPoster != null) {
-        sb.append(" to ").append(forumPoster.getDisplayName()).append(" success = ")
-            .append(String.valueOf(forumSuccess));
+      if (forumSuccess != null) {
+        sb.append(" to ").append(forumPoster.get().getDisplayName()).append(" success = ")
+            .append(forumSuccess.isDone() && !forumSuccess.isCancelled());
       }
-      if (emailSender != null) {
-        if (forumPoster != null) {
-          sb.append(" and to ");
-        } else {
-          sb.append(" to ");
-        }
-        sb.append(emailSender.getToAddress()).append(" success = ").append(String.valueOf(emailSuccess));
+      if (emailSender.isPresent()) {
+        sb.append(forumPoster.isPresent() ? " and to " : " to ");
+        sb.append(gameProperties.get(IEmailSender.RECIPIENTS)).append(" success = ").append(emailSuccess);
       }
       historyWriter.startEvent(sb.toString());
     }
-    return forumSuccess && emailSuccess;
+    return (forumSuccess == null || !forumSuccess.isCancelled()) && emailSuccess;
   }
 
   /**
@@ -160,29 +148,9 @@ public class PbemMessagePoster implements Serializable {
     return "<pre><br/>" + string.replaceAll("\n", "<br/>") + "<br/></pre>";
   }
 
-  /**
-   * Get the configured email sender.
-   *
-   * @return return an email sender or null
-   */
-  public IEmailSender getEmailSender() {
-    return emailSender;
-  }
-
-  /**
-   * Return the status string from sending the email.
-   *
-   * @return a success of failure string, or null if no email sender was configured
-   */
-  public String getEmailSendStatus() {
-    return emailSendStatus;
-  }
-
   public boolean alsoPostMoveSummary() {
-    if (forumPoster != null) {
-      return forumPoster.getAlsoPostAfterCombatMove();
-    }
-    return emailSender != null && emailSender.getAlsoPostAfterCombatMove();
+    return gameProperties.get(IForumPoster.POST_AFTER_COMBAT,
+        gameProperties.get(IEmailSender.POST_AFTER_COMBAT, false));
   }
 
   /**
@@ -190,22 +158,22 @@ public class PbemMessagePoster implements Serializable {
    * and forum (if provided). The user is first prompted to confirm they wish to perform the action before the turn is
    * posted.
    */
-  public static void postTurn(final String title, final HistoryLog historyLog, final boolean includeSaveGame,
-      final PbemMessagePoster posterPbem, final IAbstractForumPosterDelegate postingDelegate,
+  public void postTurn(final String title, final HistoryLog historyLog, final boolean includeSaveGame,
+      final IAbstractForumPosterDelegate postingDelegate,
       final TripleAFrame frame, final JComponent postButton) {
     String message = "";
-    final IForumPoster turnSummaryMsgr = posterPbem.getForumPoster();
+    final String displayName = gameProperties.get(IForumPoster.NAME, "");
     final StringBuilder sb = new StringBuilder();
-    if (turnSummaryMsgr != null) {
+    if (!displayName.isEmpty()) {
       sb.append(message).append("Post ").append(title).append(" ");
       if (includeSaveGame) {
         sb.append("and save game ");
       }
-      sb.append("to ").append(turnSummaryMsgr.getDisplayName()).append("?\n");
+      sb.append("to ").append(displayName).append("?\n");
     }
-    final IEmailSender emailSender = posterPbem.getEmailSender();
-    if (emailSender != null) {
-      sb.append("Send email to ").append(emailSender.getToAddress()).append("?\n");
+    final String opponent = gameProperties.get(IEmailSender.RECIPIENTS, "");
+    if (!opponent.isEmpty()) {
+      sb.append("Send email to ").append(opponent).append("?\n");
     }
     message = sb.toString();
     final int choice = JOptionPane.showConfirmDialog(frame, message, "Post " + title + "?",
@@ -226,20 +194,20 @@ public class PbemMessagePoster implements Serializable {
         try {
           saveGameFile = File.createTempFile("triplea", GameDataFileUtils.getExtension());
           frame.getGame().saveGame(saveGameFile);
-          posterPbem.setSaveGame(saveGameFile);
+          setSaveGame(saveGameFile);
         } catch (final Exception e) {
           postOk = false;
           log.log(Level.SEVERE, "Failed to create save game", e);
         }
-        posterPbem.setTurnSummary(historyLog.toString());
+        turnSummary = historyLog.toString();
         try {
           // forward the poster to the delegate which invokes post() on the poster
           if (postingDelegate != null) {
-            if (!postingDelegate.postTurnSummary(posterPbem, title, includeSaveGame)) {
+            if (!postingDelegate.postTurnSummary(this, title)) {
               postOk = false;
             }
           } else {
-            if (!posterPbem.post(null, title, includeSaveGame)) {
+            if (!post(null, title)) {
               postOk = false;
             }
           }
@@ -251,24 +219,21 @@ public class PbemMessagePoster implements Serializable {
           postingDelegate.setHasPostedTurnSummary(postOk);
         }
         final StringBuilder sb1 = new StringBuilder();
-        if (posterPbem.getForumPoster() != null) {
-          final String saveGameRef = posterPbem.getSaveGameRef();
-          final String turnSummaryRef = posterPbem.getTurnSummaryRef();
-          if (saveGameRef != null) {
-            sb1.append("\nSave Game : ").append(saveGameRef);
-          }
-          if (turnSummaryRef != null) {
-            sb1.append("\nSummary Text: ").append(turnSummaryRef);
+        if (gameProperties.get(IForumPoster.NAME) != null) {
+          if (this.turnSummaryRef != null) {
+            sb1.append("\nSummary Text: ").append(this.turnSummaryRef);
           }
         }
-        if (posterPbem.getEmailSender() != null) {
-          sb1.append("\nEmails: ").append(posterPbem.getEmailSendStatus());
+        if (gameProperties.get(IEmailSender.SUBJECT) != null) {
+          sb1.append("\nEmails: ").append(emailSendStatus);
         }
         historyLog.getWriter().println(sb1.toString());
         if (historyLog.isVisible()) {
           historyLog.setVisible(true);
         }
-        saveGameFile.delete();
+        if (saveGameFile != null) {
+          saveGameFile.delete();
+        }
         progressWindow.setVisible(false);
         progressWindow.removeAll();
         progressWindow.dispose();
@@ -278,15 +243,20 @@ public class PbemMessagePoster implements Serializable {
           if (postButton != null) {
             postButton.setEnabled(!finalPostOk);
           }
-          if (finalPostOk) {
-            JOptionPane.showMessageDialog(frame, finalMessage, title + " Posted",
-                JOptionPane.INFORMATION_MESSAGE);
-          } else {
-            JOptionPane.showMessageDialog(frame, finalMessage, title + " Posted",
-                JOptionPane.ERROR_MESSAGE);
-          }
+          JOptionPane.showMessageDialog(frame, finalMessage, title + " Posted",
+              finalPostOk ? JOptionPane.INFORMATION_MESSAGE : JOptionPane.ERROR_MESSAGE);
         });
       }).start();
     }
+  }
+
+  @SuppressWarnings("static-method")
+  private void readObject(@SuppressWarnings("unused") final ObjectInputStream stream) {
+    throw new UnsupportedOperationException("This class shouldn't get de-serialized!");
+  }
+
+  @SuppressWarnings("static-method")
+  private void writeObject(@SuppressWarnings("unused") final ObjectOutputStream stream) {
+    throw new UnsupportedOperationException("This class shouldn't get serialized!");
   }
 }
